PROJECT_NN/bug_triage_assistant_complete/requirements.txt
flask
crewai
litellm
google-generativeai
python-dotenv
